Fix for missing file

+++ protocol-common/src/main/java/org/kendar/storage/FileStorageRepository.java
@@ -650,6 +650,9 @@ public class FileStorageRepository extends StorageRepository {
     public void deleteFile(String... path) {
         var realPath = buildRealPath(path)+".json";
         try {
+            if(!fileExists(realPath)){
+                return;
+            }
             Files.deleteIfExists(Path.of(realPath));
         } catch (IOException e) {
             throw new RuntimeException(e);
@@ -659,6 +662,9 @@ public class FileStorageRepository extends StorageRepository {
     public String readFile(String ... path) {
         var realPath = buildRealPath(path)+".json";
         try {
+            if(!fileExists(realPath)){
+                return null;
+            }
             return getFileContent(Path.of(realPath));
         } catch (IOException e) {
             throw new RuntimeException(e);
